+++ src/mma/preprocess/__init__.py
@@ -5,11 +5,13 @@ from mma.preprocess.build_processed import (
     build_processed_batch,
     write_processed_manifest,
 )
+from mma.preprocess.load_processed import load_processed_items
 from mma.preprocess.pair_images_excel import pair_images_with_excel
 
 __all__ = [
     "assign_image_ids",
     "build_processed_batch",
+    "load_processed_items",
     "pair_images_with_excel",
     "write_processed_manifest",
 ]